all server actions

+++ src/server/actions.ts
@@ -1,19 +1,107 @@
 'use server';
 
+import bcrypt from 'bcrypt';
 import prisma from "@/lib/db";
 import { Pet } from "@prisma/client";
 import { sleep } from "@/lib/utils";
 import { revalidatePath } from "next/cache";
 import { PetNew } from "@/lib/type";
-import { petFormSchema, petIdSchema } from "@/lib/schema";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/schema";
+import { redirect } from "next/navigation";
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { Prisma } from "@prisma/client";
+import { AuthError } from "next-auth";
+import { auth, signIn, signOut } from "@/lib/auth-no-edge";
 
-export async function addPet(newPet: unknown) {
+const stripe = require("stripe")(process.env.STRIPE_SECRET_KEY);
+// --- user actions ---
 
-    await sleep(1000);
+export async function logIn(prevState: unknown, formData: unknown) {
+    if (!(formData instanceof FormData)) {
+        return {
+            message: "Invalid form data.",
+        };
+    }
 
-    const validatedPet = petFormSchema.safeParse(newPet);
+    try {
+        await signIn("credentials", formData);
+    } catch (error) {
+        if (error instanceof AuthError) {
+            switch (error.type) {
+                case "CredentialsSignin": {
+                    return {
+                        message: "Invalid credentials.",
+                    };
+                }
+                default: {
+                    return {
+                        message: "Error. Could not sign in.",
+                    };
+                }
+            }
+        }
+
+        throw error; // nextjs redirects throws error, so we need to rethrow it
+    }
+}
+
+export async function signUp(prevState: unknown, formData: unknown) {
+    // check if formData is a FormData type
+    if (!(formData instanceof FormData)) {
+        return {
+            message: "Invalid form data.",
+        };
+    }
+
+    // convert formData to a plain object
+    const formDataEntries = Object.fromEntries(formData.entries());
+
+    // validation
+    const validatedFormData = authSchema.safeParse(formDataEntries);
+    if (!validatedFormData.success) {
+        return {
+            message: "Invalid form data.",
+        };
+    }
+
+    const { email, password } = validatedFormData.data;
+    const hashedPassword = await bcrypt.hash(password, 10);
+    try {
+        await prisma.user.create({
+            data: {
+                email,
+                hashedPassword,
+            },
+        });
+    } catch (error) {
+        if (error instanceof Prisma.PrismaClientKnownRequestError) {
+            if (error.code === "P2002") {
+                return {
+                    message: "Email already exists.",
+                };
+            }
+        }
+
+        return {
+            message: "Could not create user.",
+        };
+    }
+
+    await signIn("credentials", formData);
+}
+
+export async function logOut() {
+    await signOut({ redirectTo: "/" });
+}
+
+// --- pet actions ---
+
+
+export async function addPet(pet: unknown) {
+    const session = await checkAuth();
+
+    const validatedPet = petFormSchema.safeParse(pet);
     if (!validatedPet.success) {
-        console.log(validatedPet.error);
         return {
             message: "Invalid pet data.",
         };
@@ -21,7 +109,14 @@ export async function addPet(newPet: unknown) {
 
     try {
         await prisma.pet.create({
-            data: validatedPet.data
+            data: {
+                ...validatedPet.data,
+                user: {
+                    connect: {
+                        id: session.user.id,
+                    },
+                },
+            },
         });
     } catch (error) {
         console.log(error);
@@ -33,9 +128,9 @@ export async function addPet(newPet: unknown) {
     revalidatePath("/app", "layout");
 }
 
-
 export async function editPet(petId: unknown, newPetData: unknown) {
     // authentication check
+    const session = await checkAuth();
 
     // validation
     const validatedPetId = petIdSchema.safeParse(petId);
@@ -48,7 +143,17 @@ export async function editPet(petId: unknown, newPetData: unknown) {
     }
 
     // authorization check
-
+    const pet = await getPetById(validatedPetId.data);
+    if (!pet) {
+        return {
+            message: "Pet not found.",
+        };
+    }
+    if (pet.userId !== session.user.id) {
+        return {
+            message: "Not authorized.",
+        };
+    }
 
     // database mutation
     try {
@@ -67,20 +172,69 @@ export async function editPet(petId: unknown, newPetData: unknown) {
     revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: string) {
-    // await sleep(1000);
+export async function deletePet(petId: unknown) {
+    // authentication check
+    const session = await checkAuth();
 
+    // validation
+    const validatedPetId = petIdSchema.safeParse(petId);
+    if (!validatedPetId.success) {
+        return {
+            message: "Invalid pet data.",
+        };
+    }
+
+    // authorization check
+    const pet = await getPetById(validatedPetId.data);
+    if (!pet) {
+        return {
+            message: "Pet not found.",
+        };
+    }
+    if (pet.userId !== session.user.id) {
+        return {
+            message: "Not authorized.",
+        };
+    }
+
+    // database mutation
     try {
         await prisma.pet.delete({
             where: {
-                id: petId,
+                id: validatedPetId.data,
             },
         });
     } catch (error) {
         return {
-            message: "Could not delete pet"
-        }
+            message: "Could not delete pet.",
+        };
     }
 
-    revalidatePath('/app', 'layout');
+    revalidatePath("/app", "layout");
+}
+
+// --- payment actions ---
+
+export async function createCheckoutSession() {
+    // authentication check
+    const session = await checkAuth();
+
+    console.log(session.user.email);
+
+    // create checkout session
+    const checkoutSession = await stripe.checkout.sessions.create({
+        customer_email: session.user.email,
+        line_items: [
+            {
+                price: "price_1OfpJ7FIW685mC8GCahpbCed",
+                quantity: 1,
+            },
+        ],
+        mode: "payment",
+        success_url: `${process.env.CANONICAL_URL}/payment?success=true`,
+        cancel_url: `${process.env.CANONICAL_URL}/payment?cancelled=true`,
+    });
+
+    // redirect user
+    redirect(checkoutSession.url);
 }
